feat(configuration): Update environment variables and add WebSocket support
- Update .env.example with database credentials and JWT expiration settings
- Add Mail and CORS configurations for enhanced server capabilities
- Include spring-boot-starter-websocket dependency in pom.xml for real-time communication
- Modify entity IDs from Long to Integer for consistency across the application
- Implement pre-persist timestamps in DoctorRequest and MedicalRecord entities
- Enhance GraphQL exception handling and security configurations
- Refactor message-related services and repositories to support new ID types
- Introduce new messaging UI components and hooks for improved user experience
- Update tests to reflect changes in data types and messaging functionalities

+++ family_medical/src/main/java/com/example/famMedical/Entity/Conversation.java
@@ -22,7 +22,7 @@ public class Conversation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "conversation_id")
-    private Long conversationID;
+    private Integer conversationID;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "doctor_id", nullable = false)
