+++ family_medical/src/main/java/com/example/famMedical/Entity/Message.java
@@ -24,7 +24,7 @@ public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "message_id")
-    private Long messageID;
+    private Integer messageID;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "conversation_id", nullable = false)
